Move access token file to a different directory and resolve authentication issue with user role detection
